// website/src/types.ts
export type SkillCategory =
  | 'languages'
  | 'web'
  | 'backend'
  | 'datascience'
  | 'cloud'
  | 'os'
  | 'tools'
  | 'offsec'
  | 'dfir'
  | 'fullstack'
  | 'crypto'
  | string;

export interface Skill {
  id: string;
  title: string;
  category: SkillCategory;
  level: number;
  levelLabel: string;
  expYears: string;
  command: string;
  icon: string; // Remix icon class e.g. 'ri-shield-keyhole-line'
  description: string;
  tags?: string[];
  ecosystem?: string;
  codeSample?: string;
}